[#122] Separated tests appropriately and made test more robust
